+++ src/models/Volume.php
@@ -13,6 +13,7 @@ use craft\base\FsInterface;
 use craft\base\Model;
 use craft\behaviors\FieldLayoutBehavior;
 use craft\elements\Asset;
+use craft\fs\MissingFs;
 use craft\helpers\App;
 use craft\helpers\ArrayHelper;
 use craft\records\Volume as VolumeRecord;
@@ -214,7 +215,8 @@ class Volume extends Model
             }
             $fs = Craft::$app->getFs()->getFilesystemByHandle($handle);
             if (!$fs) {
-                throw new InvalidConfigException("Invalid filesystem handle: $this->_fsHandle");
+                Craft::error("Invalid filesystem handle: $this->_fsHandle for the $this->name volume.");
+                return new MissingFs(['handle' => $this->_fsHandle]);
             }
             $this->_fs = $fs;
         }